test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
hacking>=3.0,<3.1.0 # Apache-2.0

coverage!=4.4,>=4.0 # Apache-2.0
ddt>=1.0.1 # MIT
extras>=1.0.0 # MIT
fixtures>=3.0.0 # Apache-2.0/BSD
jsonschema>=2.6.0 # MIT
mock>=3.0.0 # BSD
python-subunit>=1.0.0 # Apache-2.0/BSD
statsd>=3.3.0
stestr>=1.0.0 # Apache-2.0
testrepository>=0.0.18 # Apache-2.0/BSD
testscenarios>=0.4 # Apache-2.0/BSD
testtools>=2.2.0 # MIT
doc8>=0.8.0  # Apache-2.0
Pygments>=2.2.0  # BSD license

pytest
openstacksdk

# OTCE specifics
# Func tests of OSC
tempest>=17.1.0 # Apache-2.0
# Usability of venv
flake8
